### 2025-10-03 #9135 @editorialbot
[PRE REVIEW]: Htool-DDM: A C++ library for parallel solvers and compressed linear systems.
**Submitting author:** @PierreMarchand20 (<a href="http://orcid.org/0000-0002-2522-6837">Pierre Marchand</a>)
**Repository:** https://github.com/htool-ddm/htool
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
