### 2025-10-03 12c082d
fix .bib
