+++ paper.bib
@@ -1,33 +1,37 @@
-@book{Hackbusch2016,
-  title     = {Hierarchical Matrices: Algorithms and Analysis},
-  publisher = {Springer-Verlag, Berlin},
-  year      = {2015},
+@book{Hackbusch2015HMA,
   author    = {Hackbusch, Wolfgang},
-  volume    = {49},
+  publisher = {Springer-Verlag, Berlin},
+  title     = {Hierarchical Matrices: Algorithms and Analysis},
+  year      = {2015},
   series    = {Springer Series in Computational Mathematics},
+  volume    = {49},
   doi       = {10.1007/978-3-662-47324-5},
   groups    = {Compression techniques > Hierarchical matrix > General,Compression techniques > Hierarchical matrix > H matrix,Compression techniques > Hierarchical matrix > H2 matrix,Compression techniques > Hierarchical matrix > ACA}
 }
 
-@article{Boerm2003,
+
+@article{BoermGrasedyckEtAl2003IHM,
   author    = {B{\"o}rm, Steffen and Grasedyck, Lars and Hackbusch, Wolfgang},
-  title     = {Introduction to hierarchical matrices with applications},
   journal   = {Engineering Analysis with Boundary Elements},
+  title     = {Introduction to hierarchical matrices with applications},
   year      = {2003},
-  volume    = {27},
+  month     = {may},
   number    = {5},
   pages     = {405--422},
+  volume    = {27},
+  doi       = {10.1016/S0955-7997(02)00152-2},
   groups    = {Compression techniques > Hierarchical matrix > H matrix, Compression techniques > Hierarchical matrix > H2 matrix},
   publisher = {Elsevier}
 }
 
-@book{Bebendorf2008,
-  title     = {Hierarchical matrices: A Means to Efficiently Solve Elliptic Boundary Value Problems},
-  publisher = {Springer-Verlag, Berlin},
-  year      = {2008},
+@book{Bebendorf2008HMM,
   author    = {Bebendorf, Mario},
-  volume    = {63},
+  publisher = {Springer Berlin Heidelberg},
+  title     = {Hierarchical matrices: A Means to Efficiently Solve Elliptic Boundary Value Problems},
+  year      = {2008},
   series    = {Lecture Notes in Computational Science and Engineering},
+  volume    = {63},
+  doi       = {10.1007/978-3-540-77147-0},
   groups    = {Compression techniques > Hierarchical matrix > ACA, Compression techniques > Hierarchical matrix > H matrix, Compression techniques > Hierarchical matrix > General}
 }
 
@@ -37,9 +41,11 @@
   title     = {Efficient Numerical Methods for Non-local Operators: $\mathcal{h}^2$-matrix Compression, Algorithms and Analysis (Ems Tracts in Mathematics)},
   year      = {2010},
   isbn      = {978-3-03719-091-3},
+  doi       = {10.4171/091},
   groups    = {Compression techniques > Hierarchical matrix > H2 matrix},
   url       = {https://www.amazon.com/Efficient-Numerical-Methods-Non-local-Operators/dp/3037190914?SubscriptionId=AKIAIOBINVZYXZQZ2U3A&tag=chimbori05-20&linkCode=xm2&camp=2025&creative=165953&creativeASIN=3037190914}
 }
+
 
 
 @inproceedings{JolivetHechtEtAl2013SDD,
@@ -49,7 +55,6 @@
   year      = {2013},
   publisher = {{ACM} Press},
   doi       = {10.1145/2503210.2503212},
-  file      = {:JolivetHechtEtAl2013SDD.pdf:PDF},
   groups    = {PDE > DDM, Software}
 }
 
@@ -158,6 +163,7 @@
[…]
