### 2025-10-03 #9135 @PierreMarchand20
@danielskatz thank you for the quick answer. I fixed the references, the remaining references without DOI are a paper without DOI, my PhD thesis and two other libraries (all accessible via a url except for the first).

Looking at the list of possible reviewers, we can suggest several persons working on the numerical solution of PDEs and HPC:
- mscroggs (Matthew W. Scroggs)
- victorapm (Victor Antonio Paludetto Magri)
- lukeolson (Luke Olson)
- SoilRos (Santiago Ospina De Los Ríos)
