### 2025-10-03 #9135 @danielskatz
As another minor point, regarding the affiliations, we really only need organization, city, and country. We don't need addresses beyond this, as we don't expect anyone to be sending physical mail...
