### 2025-10-08 #9135 @gkthiruvathukal
@victorapm, @lukeolson, and @SoilRos: Can you please let me know whether you can help with this JOSS submission. I need two reviewers in order to proceed with review. Your names have been suggested by the author. (I did not ask @mscroggs, who is already helping with another submission that I'm editing.)
