### 2025-10-08 #9135 @victorapm
Thanks for inviting, George. I'm busy with another JOSS review now, but I can take a look at this next month if you need. Best!
