### 2025-10-15 328d4bc
Avoid empty-sized gemv
 BLAS error: Parameter lda passed to cblas_zgemv was 0, which is invalid.
