### 2025-10-16 5b467a8
all scaling operations use blas scal
