### 2025-10-16 de7cc4b
all scaling operations use blas scal
