### 2025-11-05 htool-ddm/htool#65 @nilsfriess
[joss-review] Minimal CMake example is missing some dependencies
Hi,
while trying to build and install htool as described in the docs (which worked without problems), I also wanted to try out the CMake example. It lists these packages as dependencies:

```cmake
find_package(MPI REQUIRED)
find_package(BLAS REQUIRED)
```
I had to additionally specify
```cmake
find_package(LAPACK REQUIRED)
find_package(OpenMP REQUIRED)
```
It seems that all those packages are hard dependencies of htool, so I was wondering if it would make more sense if the HtoolConfig.cmake just does the find_package calls? Then a simple find_package(Htool REQUIRED) would be sufficient. I am not an expert in CMake, so if there are some disadvantages to this, then the example should still be updated (I'd be happy to create a PR for that but I'm not sure where the docs can even be edited, I guess they are in a separate repo?).

cc:  openjournals/joss-reviews#9279
