### 2025-11-07 htool-ddm/htool#65 @nilsfriess
The problem is that htool itself has `find_package(OpenMP)` in its CMakeLists.txt, so when I build htool and it finds OpenMP, then the `Htool::htool` target in HtoolTargets.cmake will link against `OpenMP::OpenMP_CXX`. So at this point OpenMP is not an optional dependency anymore and I need to have `find_package(OpenMP)` (ideally `find_package(OpenMP REQUIRED)`) in my CMakeLists.txt if I want to use `Htool::htool`.

So I think this

> That being said, I checked how to put these find_package calls into the CMake target and it seems possible to add them in here.

would be the best solution. And you could ship the `FindHPDDM.cmake` with the other CMake files and use it there as well.

But I think clarifying the example is also fine.
