### 2025-11-07 htool-ddm/htool#66 @nilsfriess
[joss-review] Some minor errors in the citations
Hi,
here is a list of a few small errors in the list of references of the joss draft:

- The citation for the PETSc users manual says it is for Revision 3.24, but the doi is for 3.23
- In the Börm (2010) citation:
  1. The author's name has a typo: Borm vs. Börm
  2. I am not sure if "(ems tracts in mathematics)" should be part of the title
  3. The H^2 in the title is typeset in upper case in the original title but is lower case in the citation.
- Springer is cited in different ways: as `Springer Berlin Heidelberg`, and as `Springer Verlag, Heidelberg`
- In the Hebeker (1990) reference, the editors are named, in the Ambikasaran & Sankaran (2019) reference, they are not.
- The ending page number in Hecht (2012) is incorrect (should be 265 instead of 266)
- The ending page number in Heuer (1996) is incorrect (should 361 instead of 360)
- In the Iske, A., Borne, S. L., & Wende, M. (2017). H citation, `Borne, S. L.` should be `Le Borne, S.`
- The Jolivet, P., Hecht, F., Nataf, F., & Prud’homme, C. (2013). S citation does not have page numbers
- In the Zampini, S., Boukaram, W., Turkiyyah, G., Knio, O., & Keyes, D. (2022) reference, the article number should be 31 instead of 3

cc: openjournals/joss-reviews#9279
