### 2025-11-07 2c33a09
fix .bib
