+++ paper.bib
@@ -1,6 +1,6 @@
 @book{Hackbusch2015HMA,
   author    = {Hackbusch, Wolfgang},
-  publisher = {Springer-Verlag, Berlin},
+  publisher = {Springer Berlin Heidelberg},
   title     = {Hierarchical Matrices: Algorithms and Analysis},
   year      = {2015},
   series    = {Springer Series in Computational Mathematics},
@@ -21,7 +21,7 @@
   volume    = {27},
   doi       = {10.1016/S0955-7997(02)00152-2},
   groups    = {Compression techniques > Hierarchical matrix > H matrix, Compression techniques > Hierarchical matrix > H2 matrix},
-  publisher = {Elsevier}
+  publisher = {Elsevier BV}
 }
 
 @book{Bebendorf2008HMM,
@@ -36,9 +36,9 @@
 }
 
 @book{Borm2010ENM,
-  author    = {Steffen Borm},
+  author    = {Steffen B{\"o}rm},
   publisher = {European Mathematical Society},
-  title     = {Efficient Numerical Methods for Non-local Operators: $\mathcal{h}^2$-matrix Compression, Algorithms and Analysis (Ems Tracts in Mathematics)},
+  title     = {Efficient Numerical Methods for Non-local Operators: $\mathcal{H}^2$-matrix Compression, Algorithms and Analysis},
   year      = {2010},
   isbn      = {978-3-03719-091-3},
   doi       = {10.4171/091},
@@ -49,13 +49,18 @@
 
 
 @inproceedings{JolivetHechtEtAl2013SDD,
-  author    = {Pierre Jolivet and Fr{\'{e}}d{\'{e}}ric Hecht and Fr{\'{e}}d{\'{e}}ric Nataf and Christophe Prud'homme},
-  booktitle = {Proceedings of the International Conference for High Performance Computing, Networking, Storage and Analysis on - {SC} {\textquotesingle}13},
-  title     = {Scalable domain decomposition preconditioners for heterogeneous elliptic problems},
-  year      = {2013},
-  publisher = {{ACM} Press},
-  doi       = {10.1145/2503210.2503212},
-  groups    = {PDE > DDM, Software}
+  author     = {Pierre Jolivet and Fr{\'{e}}d{\'{e}}ric Hecht and Fr{\'{e}}d{\'{e}}ric Nataf and Christophe Prud{\textquotesingle}homme},
+  booktitle  = {Proceedings of the International Conference on High Performance Computing, Networking, Storage and Analysis},
+  title      = {Scalable domain decomposition preconditioners for heterogeneous elliptic problems},
+  year       = {2013},
+  month      = nov,
+  pages      = {1--11},
+  publisher  = {ACM},
+  series     = {SC13},
+  collection = {SC13},
+  doi        = {10.1145/2503210.2503212},
+  file       = {:JolivetHechtEtAl2013SDD.pdf:PDF},
+  groups     = {PDE > DDM, Software}
 }
 
 
@@ -77,17 +82,18 @@
   url       = {https://github.com/jeromerobert/hmat-oss}
 }
 
-@article{Ambikasaran2019,
-  doi       = {10.21105/joss.01167},
-  url       = {https://doi.org/10.21105/joss.01167},
+@article{AmbikasaranSinghEtAl2019HLH,
+  author    = {Ambikasaran, Sivaram and Singh, Karan and Sankaran, Shyam},
+  journal   = {Journal of Open Source Software},
+  title     = {HODLRlib: A Library for Hierarchical Matrices},
   year      = {2019},
-  publisher = {The Open Journal},
-  volume    = {4},
+  issn      = {2475-9066},
+  month     = feb,
   number    = {34},
[…]
