### 2025-11-07 htool-ddm/htool#66 @PierreMarchand20
Thank you for your feedback. I have checked all references querying the .bib via doi to fix these issues. It fixed most of the issues (but some are related to the metadata from the publisher, that I fixed anyway), and it fixed some other typos. See 2c33a0900426e0a49579c688338f362f89abc57e

But:
- for Hebeker (1990) with editors vs Ambikasaran & Sankaran (2019) that do not have editors. The former is a proceeding while the latter is an article.
