### 2025-11-08 htool-ddm/htool#65 @PierreMarchand20
Ideally I would fix in `main` and rebase `joss_paper`, but I just want to make sure it would not break anything on JOSS side.
