### 2025-11-10 htool-ddm/htool#66 @nilsfriess
Thank you.

> * for Hebeker (1990) with editors vs Ambikasaran & Sankaran (2019) that do not have editors. The former is a proceeding while the latter is an article.

Sorry, I mixed up the reference, I meant `Ainsworth, M., & Glusa, C. (2018)` which is a book chapter in an edited volume but is missing editors.
