### 2025-11-10 htool-ddm/htool#67 @nilsfriess
[joss-review] A few minor spelling/ grammar comments
Here a few minor spelling/grammar issues I noticed. (Some of these are not strictly errors, but I wanted to mention them anyway; Feel free to ignore them, none of them would be blocking issues for me).

- Line 4f: Inria is spelt differently in lines 4 and 5
- Line 41: You cite "Generalized Eigenvalue problem in the Overlap" using quotes which makes it seem like a verbatim citation but the specific phrase does not appear in the cited paper.
- Line 43: [...] via *an* in-house [...]
- Line 48: [...] via *an* in-house/ via *its* in-house [...]
- Line 52f: [...] one can provide *one's* own/ the user can provide their own [...]
- Line 61: [...] *row* and *column* indexes [...]
- Line 80f: The commas should be removed.
- Line 81: [...] be noted *that* these libraries [...]

cc: https://github.com/openjournals/joss-reviews/issues/9279
