+++ paper.bib
@@ -131,6 +131,7 @@
 @inbook{AinsworthGlusa2018TEF,
   author    = {Mark Ainsworth and Christian Glusa},
   pages     = {17--57},
+  editor    = {Dick, Josef and Kuo, Frances Y. and Wo{\'{z}}niakowski, Henryk},
   publisher = {Springer International Publishing},
   title     = {Towards an Efficient Finite Element Method for the Integral Fractional Laplacian on Polygonal Domains},
   year      = {2018},
