+++ paper.md
@@ -19,7 +19,7 @@ authors:
 affiliations:
  - name: POEMS, CNRS, Inria, ENSTA, Institut Polytechnique de Paris, 91120 Palaiseau, France
    index: 1
- - name: Sorbonne Université, Université Paris Cité, CNRS, INRIA, Laboratoire Jacques-Louis Lions, LJLL, EPC ALPINES, 4 place Jussieu, Paris F-75005, France
+ - name: Sorbonne Université, Université Paris Cité, CNRS, Inria, Laboratoire Jacques-Louis Lions, LJLL, EPC ALPINES, 4 place Jussieu, Paris F-75005, France
    index: 2
  - name: Sorbonne Université, CNRS, LIP6, 75252 Paris, France
    index: 3
@@ -54,13 +54,13 @@ Beside compression, another technique to accelerate iterative linear solvers is 
 
 A class of preconditioners stemming from domain decomposition methods (DDM) are *Schwarz preconditioners*. They rely on a decomposition with overlap of the initial domain. $\mathbf{P}$ is constructed by solving well-chosen problems within each subdomain, together with a small global problem, called coarse space, that provides the necessary global information for the method to retain robustness as the number of subdomains increases. Such preconditioners have been introduced for boundary integral equations in [@Hebeker1990PSA] and analyzed in [@StephanTran1998DDA; @Heuer1996EA].
 
-One goal of `Htool-DDM` is to provide black-box DDM solvers using an adaptive coarse space called GenEO for "Generalized Eigenvalue problem in the Overlap" [@SpillaneDoleanEtAl2013ARC]. It has been adapted for specific boundary integral problems in [@MarchandClaeysEtAl2020TLP; @Marchand2020SMB]. The library also includes a default black-box matrix compression via in-house $\mathcal{H}$-matrix implementation. `Htool-DDM` is a flexible platform to use and develop new DDM preconditioners and explore compression techniques via its multiple customization points.
+One goal of `Htool-DDM` is to provide black-box DDM solvers using an adaptive coarse space called GenEO for "Generalized Eigenproblems problem in the Overlap" [@SpillaneDoleanEtAl2013ARC]. It has been adapted for specific boundary integral problems in [@MarchandClaeysEtAl2020TLP; @Marchand2020SMB]. The library also includes a default black-box matrix compression via an in-house $\mathcal{H}$-matrix implementation. `Htool-DDM` is a flexible platform to use and develop new DDM preconditioners and explore compression techniques via its multiple customization points.
 
 # Statement of need
 
-`Htool-DDM` is a lightweight C++ library that provides an easy-to-use interface for distributed iterative solvers and standard black-box matrix compression techniques via in-house $\mathcal{H}$-matrix implementation. Its goal is to provide DDM preconditioners for dense/compressed linear systems.
+`Htool-DDM` is a lightweight C++ library that provides an easy-to-use interface for distributed iterative solvers and standard black-box matrix compression techniques via an in-house $\mathcal{H}$-matrix implementation. Its goal is to provide DDM preconditioners for dense/compressed linear systems.
 
-It has many customization points to support research on efficient DDM preconditioners and compression techniques. For example, one can provide its own compression algorithm, or customize the default hierarchical compression. Via its interface with HPDDM from [@JolivetHechtEtAl2013SDD], it is also a flexible tool to test various iterative solvers and preconditioners, where local and/or global problems associated with DDM preconditioners can be tailored to the problem at hand.
[…]
