### 2025-11-10 e7f9801
Fix checks for correct number of arguments
