### 2025-11-10 htool-ddm/htool#69 @prj-
We need to sort this out, I've the same issue on a machine where I don't have root privileges, so it's not as easy as removing the faulty package. But yeah, not really an Htool issue per say.
