+++ examples/CMakeLists.txt
@@ -10,13 +10,17 @@ add_executable(Use_hmatrix use_hmatrix.cpp)
 target_link_libraries(Use_hmatrix htool)
 add_dependencies(build-examples Use_hmatrix)
 
-add_executable(Use_distributed_operator use_distributed_operator.cpp)
-target_link_libraries(Use_distributed_operator htool)
-add_dependencies(build-examples Use_distributed_operator)
+if(MPI_FOUND)
+    add_executable(Use_distributed_operator use_distributed_operator.cpp)
+    target_link_libraries(Use_distributed_operator htool)
+    add_dependencies(build-examples Use_distributed_operator)
+endif()
 
-add_executable(Use_ddm_solver use_ddm_solver.cpp)
-target_link_libraries(Use_ddm_solver htool)
-add_dependencies(build-examples Use_ddm_solver)
+if(MPI_FOUND AND HPDDM_FOUND)
+    add_executable(Use_ddm_solver use_ddm_solver.cpp)
+    target_link_libraries(Use_ddm_solver htool)
+    add_dependencies(build-examples Use_ddm_solver)
+endif()
 
 add_executable(Compression_comparison compression_comparison.cpp)
 target_link_libraries(Compression_comparison htool)
