+++ cmake/HtoolConfig.cmake.in
@@ -1,4 +1,10 @@
 @PACKAGE_INIT@
 
+include(CMakeFindDependencyMacro)
+find_package(BLAS REQUIRED)
+find_package(LAPACK)
+find_package(OpenMP)
+find_package(MPI)
+
 include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
 check_required_components("@PROJECT_NAME@")
