### 2025-11-10 htool-ddm/htool#69 @prj-
> Could you try with https://github.com/htool-ddm/htool/pull/70 if it works for you?

You'll have to reinstall the faulty package first.
