### 2025-11-28 htool-ddm/htool#71 @vijaysm
[joss-review] Issues building with just MPI on OSX
On a Macbook, I cloned the repository, and simply ran the following to configure a bare build.

```
mkdir build && cd build
cmake -DCMAKE_CXX_COMPILER=mpicxx ..
cd examples && make all
```

This generates an error when building wrapper_hpddm. This is expected as CMake didn't find HPDDM during configuration. So its misleading why the example is getting built at all. This is inconsistent and confusing.

```
$HOME/htool/include/htool/solvers/../wrappers/wrapper_hpddm.hpp:32:10: fatal error: 'HPDDM.hpp' file not found
   32 | #include <HPDDM.hpp>
      |          ^~~~~~~~~~~
```

I could run `make -k` and get past the failure to build all other examples.

Edit: I found the [documentation link](https://htool-ddm.pages.math.cnrs.fr/cpp_api/quickstart.html#quickstart) that mentions that DDM solvers require MPI, HPDDM, and LAPACK. However, without HPDDM, shouldn't the example be disabled?
