### 2025-11-28 htool-ddm/htool#65 @PierreMarchand20
This is fixed in #70. I will rebase `joss_paper` once this is merged
