### 2025-11-28 htool-ddm/htool#71 @PierreMarchand20
Thank you for your feedback. This is fixed in #70 where tests and examples depending on HPDDM are defined only if HPDDM is found.
