### 2025-11-28 a9f8a78
improve examples
