+++ examples/use_distributed_operator.cpp
@@ -64,16 +64,16 @@ int main(int argc, char *argv[]) {
     MPI_Comm_size(MPI_COMM_WORLD, &sizeWorld);
 
     // Check the number of parameters
-    if (argc != 2) {
+    if (argc > 2) {
         // Tell the user how to run the program
-        cerr << "Usage: " << argv[0] << " outputpath" << endl;
+        cerr << "Usage: " << argv[0] << " output_folder" << endl;
         /* "Usage messages" are a conventional way of telling the user
          * how to run a program if they enter the command incorrectly.
          */
         return 1;
     }
 
-    std::string outputpath = argv[1];
+    std::string output_folder = argc == 2 ? argv[1] : "./";
 
     // Geometry
     const int number_points        = 10000;
@@ -81,11 +81,12 @@ int main(int argc, char *argv[]) {
     const int number_of_partitions = sizeWorld;
     const int number_of_children   = 8;
     vector<double> coordinates(spatial_dimension * number_points);
-    create_sphere(number_points, coordinates.data());
+    create_rotated_ellipse(3, 4., 1., 0., 0., number_points, coordinates.data()); // 2d ellipse in 3d
 
     // Clustering
     ClusterTreeBuilder<double> recursive_build_strategy;
     recursive_build_strategy.set_maximal_leaf_size(100);
+    recursive_build_strategy.set_partitioning_strategy(std::make_shared<Partitioning_N<double, ComputeLargestExtent<double>, RegularSplitting<double>>>()); // this is actually the default choice
     Cluster<double> cluster = recursive_build_strategy.create_cluster_tree(number_points, spatial_dimension, coordinates.data(), number_of_children, number_of_partitions);
 
     // HMatrix parameters
@@ -122,7 +123,7 @@ int main(int argc, char *argv[]) {
         std::cout << "Information about hmatrices accross all processors\n";
     }
     print_distributed_hmatrix_information(local_hmatrix, std::cout, MPI_COMM_WORLD);
-    save_leaves_with_rank(local_hmatrix, outputpath + "/local_hmatrix_" + std::to_string(rankWorld));
+    save_leaves_with_rank(local_hmatrix, output_folder + "/local_hmatrix_" + std::to_string(rankWorld));
 
     // Finalize the MPI environment.
     MPI_Finalize();
