### 2025-11-28 0559e73
improve examples
