+++ examples/compression_comparison.sh
@@ -11,13 +11,13 @@ make Compression_comparison
 mkdir -p ../output/examples/compression_comparison
 
 # Arguments
-outputpath=../output/examples/compression_comparison/
+outputfolder=../output/examples/compression_comparison/
 distances=(1 2 3)
 
 # Run
 for distance in "${distances[@]}"
 do
-    ./examples/compression_comparison ${distance} compression_comparison_${distance}.csv ${outputpath}
+    ./examples/compression_comparison ${distance} compression_comparison_${distance}.csv ${outputfolder}
 done
 
 
