### 2025-12-01 #9279 @PierreMarchand20
@vijaysm Thank you for your feedbacks.

- concerning htool-ddm/htool#71 and htool-ddm/htool#72, I hope to have addressed them (see my answers in these issues in the main repository). Do not hesitate to reopen them if you wish additional changes.
- for the Python interface, the line [here](https://github.com/htool-ddm/htool_python/blob/18eb16c20d6731cbd9c8f579158958cd30fcc1a3/CMakeLists.txt#L73) should check that you have `mpi4py` if you have `MPI`. I am not sure why this is not triggered though. You could open an issue on the python interface repository https://github.com/htool-ddm/htool_python and share the output `pip install --verbose .`. This should print the cmake output.
- I have added documentation on how to build and run tests. See [here](https://htool-ddm.pages.math.cnrs.fr/cpp_api/developer_guide.html) and [here](https://htool-ddm.pages.math.cnrs.fr/python_api/developer_guide.html).
- For contributing, the workflow is standard on github (create a PR and we discuss it), I am not sure what needs to be clarified.
