### 2025-12-03 #9279 @vijaysm
Just a quick comment. For the contributions, I was expecting something like [this](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) as a base template. You can tune it and add more details as needed for your workflows, or for standard ones, just keep the base template. It's a set of guidelines that are useful as repositories and processes evolve. Hence, my suggestion. @PierreMarchand20
