### 2025-12-03 #9279 @PierreMarchand20
@vijaysm  This is added in Htool-DDM/htool#74 and will be merged shortly.

Note that all the changes so far have been made in `main`. I do not know how the review process ends, but I would like to rebase the branch `joss_paper` and squash all the commits once this is done.

And the current article proof is not updated with my last corrections. I will try the command I could use in the preview to regenerate it.
