### 2025-12-09 #9279 @vijaysm
I have completed my review, and I am happy with the responses and updates @gkthiruvathukal . There is one pending PR that needs to be merged ([PR #74](https://github.com/htool-ddm/htool/pull/74)). Other than that, all the other critical issues I had raised earlier have been addressed. The submission may be ready for final processing.

All the best @PierreMarchand20
