### 2025-12-09 #9279 @vijaysm
**Just a note:** I had hoped to test the Htool–PETSc linkage and run a few examples, but I haven’t had the time, and the rest of my December schedule makes it unlikely that I'll get around to it. Rather than delaying this review, I have submitted my final comments above. If I encounter any issues during further testing, I will open an issue in the Htool-DDM repository directly.
