### 2025-12-09 #9279 @PierreMarchand20
@vijaysm htool-DDM/htool#74 is merged, thank you for your review!
