### 2025-12-10 #9279 @PierreMarchand20
I have rebased and squashed `joss_paper` onto the last commit of `main`.
