### 2026-01-02 #9279 @PierreMarchand20
Hi @gkthiruvathukal, I noticed the figures were actually missing (I only had them locally). I pushed them and rebased `joss_paper` onto the last commit of main. Is it ok for you?

I will proceed with creating the archive as soon as you confirmed it is ok and there is no need for further reviews after this last change.

I can already confirm that there is no special character in the title `Htool-DDM: A C++ library for parallel solvers and compressed linear systems.`
