+++ .github/workflows/release.yml
@@ -25,17 +25,11 @@ jobs:
             echo "notes<<EOF" >> $GITHUB_OUTPUT
             echo "$NOTES" >> $GITHUB_OUTPUT
             echo "EOF" >> $GITHUB_OUTPUT
-            echo "TAG=$TAG"
-            echo "EXTRACTED NOTES:"
-            echo "$NOTES"
 
       - name: Create GitHub Release
-        uses: actions/create-release@v1
-        with:
-          tag_name: ${{ github.ref_name }}      # tag that was pushed
-          release_name: Release ${{ github.ref_name }}
-          body: ${{ steps.extract_notes.outputs.notes }}
-          draft: false
-          prerelease: false
+        run: |
+          gh release create ${{ github.ref_name }} \
+          --title "Release ${{ github.ref_name }}" \
+          --notes "$(cat RELEASE_NOTES.md)"
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
