+++ .github/workflows/release.yml
@@ -17,19 +17,19 @@ jobs:
         id: extract_notes
         run: |
             TAG="${GITHUB_REF_NAME#v}"   # remove leading 'v' if your tag is 'v1.0.0'
-            NOTES=$(awk -v tag="$TAG" '
+            awk -v tag="$TAG" '
                 /^## \[.*\]/ {in_section=0}               # Reset at every header
                 $0 ~ "^## \\[" tag "\\]" {in_section=1; next}  # Start when we find the tag
                 in_section {print}
-            ' CHANGELOG.md)
-            echo "notes<<EOF" >> $GITHUB_OUTPUT
-            echo "$NOTES" >> $GITHUB_OUTPUT
-            echo "EOF" >> $GITHUB_OUTPUT
+            ' CHANGELOG.md > RELEASE_NOTES.md
+            
+            # Just in case, trim trailing empty lines
+            sed -i '/^[[:space:]]*$/d' RELEASE_NOTES.md
 
       - name: Create GitHub Release
         run: |
           gh release create "${GITHUB_REF_NAME}" \
             --title "Release ${GITHUB_REF_NAME}" \
-            --notes "${{ steps.extract_notes.outputs.notes }}"
+            --notes RELEASE_NOTES.md
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
