### 2026-01-02 756a901
add release.yml workflow
