### 2026-01-02 78f58c9
add workflow_dispatch to release workflow
