+++ .github/workflows/release.yml
@@ -4,6 +4,11 @@ on:
   push:
     tags:
       - 'v*'  # triggers on any tag starting with 'v'
+  workflow_dispatch:
+    inputs:
+      tag:
+        description: "Tag to release"
+        required: false
 
 jobs:
   release:
