### 2026-01-02 f5cf16b
add CITATION.cff
