### 2026-02-05 fcd052d
add quadratures
