+++ include/htool/quadratures/gauss_legendre.hpp
@@ -1,0 +1,288 @@
+#ifndef HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
+#define HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
+#include "quadratures.hpp"
+
+namespace htool {
+
+template <typename T>
+struct GaussLegendreRule {
+    int degree;
+    int n_points;
+    const QuadPoint<double, 1> *points;
+};
+
+constexpr QuadPoint<double, 1> gl_1[] = {
+    {0.0, 2.0}};
+
+constexpr QuadPoint<double, 1> gl_2[] = {
+    {-0.5773502691896257, 1.0},
+    {0.5773502691896257, 1.0}};
+
+constexpr QuadPoint<double, 1> gl_3[] = {
+    {-0.7745966692414834, 0.5555555555555556},
+    {0.0, 0.8888888888888888},
+    {0.7745966692414834, 0.5555555555555556}};
+
+constexpr QuadPoint<double, 1> gl_4[] = {
+    {-0.8611363115940526, 0.3478548451374539},
+    {-0.3399810435848563, 0.6521451548625461},
+    {0.3399810435848563, 0.6521451548625461},
+    {0.8611363115940526, 0.3478548451374539}};
+
+constexpr QuadPoint<double, 1> gl_5[] = {
+    {-0.9061798459386640, 0.2369268850561891},
+    {-0.5384693101056831, 0.4786286704993665},
+    {0.0, 0.5688888888888889},
+    {0.5384693101056831, 0.4786286704993665},
+    {0.9061798459386640, 0.2369268850561891}};
+
+constexpr QuadPoint<double, 1> gl_6[] = {
+    {-0.9324695142031521, 0.1713244923791704},
+    {-0.6612093864662645, 0.3607615730481386},
+    {-0.2386191860831969, 0.4679139345726910},
+    {0.2386191860831969, 0.4679139345726910},
+    {0.6612093864662645, 0.3607615730481386},
+    {0.9324695142031521, 0.1713244923791704}};
+
+constexpr QuadPoint<double, 1> gl_7[] = {
+    {-0.9491079123427585, 0.1294849661688697},
+    {-0.7415311855993945, 0.2797053914892766},
+    {-0.4058451513773972, 0.3818300505051189},
+    {0.0, 0.4179591836734694},
+    {0.4058451513773972, 0.3818300505051189},
+    {0.7415311855993945, 0.2797053914892766},
+    {0.9491079123427585, 0.1294849661688697}};
+
+constexpr QuadPoint<double, 1> gl_8[] = {
+    {-0.9602898564975363, 0.1012285362903763},
+    {-0.7966664774136267, 0.2223810344533745},
+    {-0.5255324099163290, 0.3137066458778873},
+    {-0.1834346424956498, 0.3626837833783620},
+    {0.1834346424956498, 0.3626837833783620},
+    {0.5255324099163290, 0.3137066458778873},
+    {0.7966664774136267, 0.2223810344533745},
+    {0.9602898564975363, 0.1012285362903763}};
+
+constexpr QuadPoint<double, 1> gl_9[] = {
+    {-0.9681602395076261, 0.0812743883615744},
+    {-0.8360311073266358, 0.1806481606948574},
+    {-0.6133714327005904, 0.2606106964029354},
+    {-0.3242534234038089, 0.3123470770400029},
+    {0.0, 0.3302393550012598},
+    {0.3242534234038089, 0.3123470770400029},
+    {0.6133714327005904, 0.2606106964029354},
+    {0.8360311073266358, 0.1806481606948574},
+    {0.9681602395076261, 0.0812743883615744}};
+
+constexpr QuadPoint<double, 1> gl_10[] = {
[…]
