+++ include/htool/quadratures/dunavant.hpp
@@ -1,0 +1,125 @@
+#ifndef HTOOL_QUADRATURES_DUNAVANT_HPP
+#define HTOOL_QUADRATURES_DUNAVANT_HPP
+
+#include "quadratures.hpp"
+
+namespace htool {
+
+template <typename T>
+struct DunavantRule {
+    int degree;
+    int n_points;
+    const QuadPoint<T, 2> *points;
+};
+
+// ---------- Degree 1 (1 point) ----------
+constexpr QuadPoint<double, 2> dunavant_deg1[] = {
+    {1.0 / 3.0, 1.0 / 3.0, 0.5}};
+
+// ---------- Degree 2 (3 points) ----------
+constexpr QuadPoint<double, 2> dunavant_deg2[] = {
+    {1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0},
+    {2.0 / 3.0, 1.0 / 6.0, 1.0 / 6.0},
+    {1.0 / 6.0, 2.0 / 3.0, 1.0 / 6.0}};
+
+// ---------- Degree 3 (4 points) ----------
+constexpr QuadPoint<double, 2> dunavant_deg3[] = {
+    {1.0 / 3.0, 1.0 / 3.0, -27.0 / 96.0},
+    {0.2, 0.2, 25.0 / 96.0},
+    {0.6, 0.2, 25.0 / 96.0},
+    {0.2, 0.6, 25.0 / 96.0}};
+
+// ---------- Degree 4 (6 points) ----------
+constexpr QuadPoint<double, 2> dunavant_deg4[] = {
+    {0.445948490915965, 0.445948490915965, 0.111690794839005},
+    {0.108103018168070, 0.445948490915965, 0.111690794839005},
+    {0.445948490915965, 0.108103018168070, 0.111690794839005},
+    {0.091576213509771, 0.091576213509771, 0.054975871827661},
+    {0.816847572980459, 0.091576213509771, 0.054975871827661},
+    {0.091576213509771, 0.816847572980459, 0.054975871827661}};
+
+// ---------- Degree 5 (7 points) ----------
+constexpr QuadPoint<double, 2> dunavant_deg5[] = {
+    {1.0 / 3.0, 1.0 / 3.0, 0.1125},
+    {0.470142064105115, 0.470142064105115, 0.066197076394253},
+    {0.059715871789770, 0.470142064105115, 0.066197076394253},
+    {0.470142064105115, 0.059715871789770, 0.066197076394253},
+    {0.101286507323456, 0.101286507323456, 0.062969590272413},
+    {0.797426985353087, 0.101286507323456, 0.062969590272413},
+    {0.101286507323456, 0.797426985353087, 0.062969590272413}};
+
+// ---------- Degree 6 (12 points) ----------
+constexpr QuadPoint<double, 2> dunavant_deg6[] = {
+    {0.249286745170910, 0.249286745170910, 0.058393137863189},
+    {0.501426509658179, 0.249286745170910, 0.058393137863189},
+    {0.249286745170910, 0.501426509658179, 0.058393137863189},
+    {0.063089014491502, 0.063089014491502, 0.025422453185103},
+    {0.873821971016996, 0.063089014491502, 0.025422453185103},
+    {0.063089014491502, 0.873821971016996, 0.025422453185103},
+    {0.310352451033785, 0.636502499121399, 0.041425537809187},
+    {0.636502499121399, 0.053145049844816, 0.041425537809187},
+    {0.053145049844816, 0.310352451033785, 0.041425537809187},
+    {0.636502499121399, 0.310352451033785, 0.041425537809187},
+    {0.310352451033785, 0.053145049844816, 0.041425537809187},
[…]
