+++ include/htool/quadratures/gauss_legendre.hpp
@@ -1,287 +1,452 @@
 #ifndef HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
 #define HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
-#include "quadratures.hpp"
 
+#include <array>
+#include <cstddef>
 namespace htool {
 
 template <typename T>
 struct GaussLegendreRule {
-    int degree;
-    int n_points;
-    const QuadPoint<double, 1> *points;
+    std::size_t degree;
+    std::size_t nb_points;
+    const T *x;
+    const T *w;
 };
 
-constexpr QuadPoint<double, 1> gl_1[] = {
-    {0.0, 2.0}};
+// order 1
+constexpr double gl1_x[] = {0.0};
+constexpr double gl1_w[] = {2.0};
 
-constexpr QuadPoint<double, 1> gl_2[] = {
-    {-0.5773502691896257, 1.0},
-    {0.5773502691896257, 1.0}};
+// order 2
+constexpr double gl2_x[] = {
+    -0.5773502691896257,
+    0.5773502691896257};
+constexpr double gl2_w[] = {
+    1.0,
+    1.0};
 
-constexpr QuadPoint<double, 1> gl_3[] = {
-    {-0.7745966692414834, 0.5555555555555556},
-    {0.0, 0.8888888888888888},
-    {0.7745966692414834, 0.5555555555555556}};
+// order 3
+constexpr double gl3_x[] = {
+    -0.7745966692414834,
+    0.0,
+    0.7745966692414834};
+constexpr double gl3_w[] = {
+    0.5555555555555556,
+    0.8888888888888888,
+    0.5555555555555556};
 
-constexpr QuadPoint<double, 1> gl_4[] = {
-    {-0.8611363115940526, 0.3478548451374539},
-    {-0.3399810435848563, 0.6521451548625461},
-    {0.3399810435848563, 0.6521451548625461},
-    {0.8611363115940526, 0.3478548451374539}};
+// -------- order 4 --------
+constexpr double gl4_x[] = {
+    -0.8611363115940526,
+    -0.3399810435848563,
+    0.3399810435848563,
+    0.8611363115940526};
 
-constexpr QuadPoint<double, 1> gl_5[] = {
-    {-0.9061798459386640, 0.2369268850561891},
-    {-0.5384693101056831, 0.4786286704993665},
-    {0.0, 0.5688888888888889},
-    {0.5384693101056831, 0.4786286704993665},
-    {0.9061798459386640, 0.2369268850561891}};
+constexpr double gl4_w[] = {
+    0.3478548451374539,
+    0.6521451548625461,
+    0.6521451548625461,
+    0.3478548451374539};
 
-constexpr QuadPoint<double, 1> gl_6[] = {
-    {-0.9324695142031521, 0.1713244923791704},
-    {-0.6612093864662645, 0.3607615730481386},
-    {-0.2386191860831969, 0.4679139345726910},
-    {0.2386191860831969, 0.4679139345726910},
-    {0.6612093864662645, 0.3607615730481386},
-    {0.9324695142031521, 0.1713244923791704}};
+// ---------- order 5 ----------
+constexpr double gl5_x[] = {
+    -0.9061798459386640,
+    -0.5384693101056831,
+    0.0,
+    0.5384693101056831,
+    0.9061798459386640};
+constexpr double gl5_w[] = {
+    0.2369268850561891,
+    0.4786286704993665,
+    0.5688888888888889,
+    0.4786286704993665,
+    0.2369268850561891};
 
-constexpr QuadPoint<double, 1> gl_7[] = {
-    {-0.9491079123427585, 0.1294849661688697},
[…]
