+++ include/htool/quadratures/triangle.hpp
@@ -1,0 +1,149 @@
+#ifndef HTOOL_QUADRATURES_triangle_HPP
+#define HTOOL_QUADRATURES_triangle_HPP
+
+#include "quadrature.hpp"
+#include <cstddef>
+#include <string>
+namespace htool {
+
+template <typename T>
+struct TriangleRule {
+    inline static const std::string name = "triangle";
+    std::size_t degree;
+    std::size_t nb_points;
+    const QuadPoint<T, 2> *quad_points;
+};
+
+// ---------- Degree 1 (1 point) ----------
+// Centroid
+template <typename T>
+constexpr QuadPoint<T, 2> triangle_deg1[] = {
+    {{T(1.0) / T(3.0), T(1.0) / T(3.0)}, T(1)}};
+
+// ---------- Degree 2 (3 points) ----------
+// Hammer-Marlowe-Stroud
+template <typename T>
+constexpr QuadPoint<T, 2> triangle_deg2[] = {
+    {{T(1.0) / T(6.0), T(1.0) / T(6.0)}, T(1.0) / T(3.0)},
+    {{T(2.0) / T(3.0), T(1.0) / T(6.0)}, T(1.0) / T(3.0)},
+    {{T(1.0) / T(6.0), T(2.0) / T(3.0)}, T(1.0) / T(3.0)},
+};
+
+// ---------- Degree 3 (4 points) ----------
+// Hammer-Marlowe-Stroud
+template <typename T>
+constexpr QuadPoint<T, 2> triangle_deg3[] = {
+    {{T(1.0) / T(3.0), T(1.0) / T(3.0)}, T(-9.0) / T(16.0)},
+    {{T(0.2), T(0.2)}, T(25.0) / T(48.0)},
+    {{T(0.6), T(0.2)}, T(25.0) / T(48.0)},
+    {{T(0.2), T(0.6)}, T(25.0) / T(48.0)},
+};
+
+// ---------- Degree 4 (6 points) ----------
+// Cowper
+template <typename T>
+constexpr QuadPoint<T, 2> triangle_deg4[] = {
+    {{T(0.445948490915965), T(0.108103018168070)}, T(0.223381589678011)},
+    {{T(0.108103018168070), T(0.445948490915965)}, T(0.223381589678011)},
+    {{T(0.445948490915965), T(0.445948490915965)}, T(0.223381589678011)},
+    {{T(0.091576213509771), T(0.816847572980459)}, T(0.109951743655322)},
+    {{T(0.816847572980459), T(0.091576213509771)}, T(0.109951743655322)},
+    {{T(0.091576213509771), T(0.091576213509771)}, T(0.109951743655322)},
+};
+
+// ---------- Degree 5 (7 points) ----------
+// Tabulated values from https://github.com/zmoitier/HAdaptiveIntegration.jl based on Laurie, 1982 https://doi.org/10.1145/355993.356001.
+template <typename T>
+constexpr QuadPoint<T, 2> triangle_deg5[] = {
+    {{T(1.0) / T(3.0), T(1.0) / T(3.0)}, T(9.0) / T(40.0)},
+    {{T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265), T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867)}, T(0.1259391805448271525956839455001813336576392319122570076445103551949054307567196)},
+    {{T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867), T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265)}, T(0.1259391805448271525956839455001813336576392319122570076445103551949054307567196)},
[…]
