### 2026-02-11 b7e499b
update triangle quadratures
