+++ include/htool/quadratures/radon_laurie.hpp
@@ -1,74 +1,0 @@
-#ifndef HTOOL_QUADRATURES_RADON_LAURIE_HPP
-#define HTOOL_QUADRATURES_RADON_LAURIE_HPP
-
-// Based on the article: Laurie, 1982 https://doi.org/10.1145/355993.356001.
-// Tabulated values from https://github.com/zmoitier/HAdaptiveIntegration.jl
-
-#include "quadrature.hpp"
-#include <array>
-#include <cstddef>
-#include <string>
-namespace htool {
-
-template <typename T>
-struct RadonRule {
-    inline static const std::string name = "Radon";
-    std::size_t degree;
-    std::size_t nb_points;
-    const QuadPoint<T, 2> *quad_points;
-};
-
-template <typename T>
-struct LaurieRule {
-    inline static const std::string name = "Laurie";
-    std::size_t degree;
-    std::size_t nb_points;
-    const QuadPoint<T, 2> *quad_points;
-};
-
-template <typename T>
-constexpr QuadPoint<T, 2> radon[] = {
-    {{T(0.3333333333333333333333333333333333333333333333333333333333333333333333333333348), T(0.3333333333333333333333333333333333333333333333333333333333333333333333333333348)}, T(0.1124999999999999999999999999999999999999999999999999999999999999999999999999996)},
-    {{T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265), T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867)}, T(0.0629695902724135762978419727500906668288196159561285038222551775974527153783598)},
-    {{T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867), T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265)}, T(0.0629695902724135762978419727500906668288196159561285038222551775974527153783598)},
-    {{T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867), T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867)}, T(0.0629695902724135762978419727500906668288196159561285038222551775974527153783598)},
-    {{T(0.05971587178976982045911758097310479896829317092461039746785011751300161029171882), T(0.4701420641051150897704412095134476005158534145376948012660749412434991948541406)}, T(0.06619707639425309036882469391657599983784705071053816284441148906921395128830629)},
-    {{T(0.4701420641051150897704412095134476005158534145376948012660749412434991948541406), T(0.05971587178976982045911758097310479896829317092461039746785011751300161029171882)}, T(0.06619707639425309036882469391657599983784705071053816284441148906921395128830629)},
-    {{T(0.4701420641051150897704412095134476005158534145376948012660749412434991948541406), T(0.4701420641051150897704412095134476005158534145376948012660749412434991948541406)}, T(0.06619707639425309036882469391657599983784705071053816284441148906921395128830629)}};
-
-template <typename T>
-constexpr QuadPoint<T, 2> laurie[] = {
-    {{T(0.3333333333333333333333333333333333333333333333333333333333333333333333333333348), T(0.3333333333333333333333333333333333333333333333333333333333333333333333333333348)}, T(0.01893054560015734165415411067800806149331579616112555544752071088518874456461607)},
-    {{T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265), T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867)}, T(0.01881021270659148607215700260658383995407691889410336669994070251334830384521297)},
-    {{T(0.1012865073234563388009873619151238280555751568908766273053536301850722337172867), T(0.7974269853530873223980252761697523438888496862182467453892927396298555325654265)}, T(0.01881021270659148607215700260658383995407691889410336669994070251334830384521297)},
[…]
