### 2026-02-11 3fb3b69
update gauss legendre
