+++ include/htool/quadratures/gauss_legendre.hpp
@@ -1,6 +1,8 @@
 #ifndef HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
 #define HTOOL_QUADRATURES_GAUSS_LEGENDRE_HPP
 
+// This code is generated from a Python script calling legacy-quadpy.
+
 #include "quadrature.hpp"
 #include <cstddef>
 #include <string>
@@ -8,304 +10,595 @@ namespace htool {
 
 template <typename T>
 struct GaussLegendreRule {
-    inline static const std::string name = "GaussLegendre";
+    inline static const std::string name = "gauss_legendre";
     std::size_t degree;
     std::size_t nb_points;
-    const QuadPoint<T, 1> *quad_points;
+    const QuadPoint<T, 2> *quad_points;
 };
 
+// ---------- Degree 1 (1 points) ----------
+// Gauss-Legendre
 template <typename T>
-constexpr QuadPoint<T, 1> gl_1[] = {
-    {{T(0.0)}, T(2.0)}};
+constexpr QuadPoint<T, 2> gauss_legendre_deg1[] = {{{T(0.0)}, T(2.0)}};
 
+// ---------- Degree 3 (2 points) ----------
+// Gauss-Legendre
 template <typename T>
-constexpr QuadPoint<T, 1> gl_2[] = {
-    {{T(-0.5773502691896257)}, T(1.0)},
-    {{T(0.5773502691896257)}, T(1.0)}};
+constexpr QuadPoint<T, 2> gauss_legendre_deg3[] = {{{T(-0.5773502691896257)}, T(1.0)},
+                                                   {{T(0.5773502691896257)}, T(1.0)}};
 
+// ---------- Degree 5 (3 points) ----------
+// Gauss-Legendre
 template <typename T>
-constexpr QuadPoint<T, 1> gl_3[] = {
-    {{T{-0.7745966692414834}}, T{0.5555555555555556}},
-    {{T{0.0}}, T{0.8888888888888888}},
-    {{T{0.7745966692414834}}, T{0.5555555555555556}}};
+constexpr QuadPoint<T, 2> gauss_legendre_deg5[] = {{{T(-0.7745966692414834)}, T(0.5555555555555555)},
+                                                   {{T(0.0)}, T(0.8888888888888892)},
+                                                   {{T(0.7745966692414834)}, T(0.5555555555555555)}};
 
+// ---------- Degree 7 (4 points) ----------
+// Gauss-Legendre
 template <typename T>
-constexpr QuadPoint<T, 1> gl_4[] = {
-    {{T{-0.8611363115940526}}, T{0.3478548451374539}},
-    {{T{-0.3399810435848563}}, T{0.6521451548625461}},
-    {{T{0.3399810435848563}}, T{0.6521451548625461}},
-    {{T{0.8611363115940526}}, T{0.3478548451374539}}};
+constexpr QuadPoint<T, 2> gauss_legendre_deg7[] = {{{T(-0.8611363115940526)}, T(0.34785484513745357)},
+                                                   {{T(-0.33998104358485626)}, T(0.6521451548625466)},
+                                                   {{T(0.33998104358485626)}, T(0.6521451548625466)},
+                                                   {{T(0.8611363115940526)}, T(0.34785484513745357)}};
 
+// ---------- Degree 9 (5 points) ----------
+// Gauss-Legendre
 template <typename T>
-constexpr QuadPoint<T, 1> gl_5[] = {
-    {{T{-0.9061798459386640}}, T{0.2369268850561891}},
-    {{T{-0.5384693101056831}}, T{0.4786286704993665}},
-    {{T{0.0}}, T{0.5688888888888889}},
-    {{T{0.5384693101056831}}, T{0.4786286704993665}},
[…]
