### 2026-02-11 8ebe17f
fix eigs order in solve_EVP_3
