### 2026-02-11 21df9b0
fix eigs order in solve_EVP_3
