+++ CMakeLists.txt
@@ -5,12 +5,12 @@ cmake_minimum_required(VERSION 3.10)
 if(${CMAKE_VERSION} VERSION_LESS 3.12)
     project(
         Htool
-        VERSION 1.0.0
+        VERSION 1.0.1
         LANGUAGES CXX)
 else()
     project(
         Htool
-        VERSION 1.0.0
+        VERSION 1.0.1
         DESCRIPTION "A header only c++ library that provides Hierarchical matrices."
         HOMEPAGE_URL "https://github.com/htool-ddm/htool"
         LANGUAGES CXX)
