+++ include/htool/htool_version.hpp
@@ -3,7 +3,7 @@
 
 #define HTOOL_VERSION_MAJOR 1
 #define HTOOL_VERSION_MINOR 0
-#define HTOOL_VERSION_SUBMINOR 1
+#define HTOOL_VERSION_SUBMINOR 2
 
 #define HTOOL_VERSION_EQ(MAJOR, MINOR, SUBMINOR) \
     ((HTOOL_VERSION_MAJOR == (MAJOR)) && (HTOOL_VERSION_MINOR == (MINOR)) && (HTOOL_VERSION_SUBMINOR == (SUBMINOR)))
