### 2026-02-15 1bce94b
make std::execution optional
