### 2026-02-15 8b5e2b6
make std::execution optional
