### 2026-02-15 7007eb7
make std::execution optional
