+++ include/htool/hmatrix/lrmat/interpolation.hpp
@@ -5,67 +5,45 @@
 #include "../../wrappers/wrapper_blas.hpp"                      // for Blas
 #include "../../wrappers/wrapper_lapack.hpp"                    // for Blas
 #include "lrmat.hpp"
+#include <functional>
 #include <map>
 #include <span>
 
+#if defined(__clang__)
+#    pragma clang diagnostic push
+#    pragma clang diagnostic ignored "-Wold-style-cast"
+#elif defined(__GNUC__) || defined(__GNUG__)
+#    pragma GCC diagnostic push
+#    pragma GCC diagnostic ignored "-Wold-style-cast"
+#endif
+
+#include "theia.hpp"
+
+#if defined(__clang__)
+#    pragma clang diagnostic pop
+#elif defined(__GNUC__) || defined(__GNUG__)
+#    pragma GCC diagnostic pop
+#endif
+
 namespace htool {
 
-inline double cheb_node(int k, int L) { return std::cos((double)(2 * k + 1) / (double)(2 * L) * M_PI); }
-inline double T(double x, int k) { return cos((double)(k)*acos(x)); }
-inline void getTr(int L, double *Tr) {
-    for (int k = 0; k < L; k++) {
-        double r = cheb_node(k, L);
-        for (int i = 0; i < L; i++) {
-            (Tr + k * L)[i] = T(r, i);
-        }
-    }
-}
-inline void getTs(double u, int L, double *Tu) {
-    for (int i = 0; i < L; i++) {
-        Tu[i] = T(u, i);
-    }
-}
-inline void C1D(double *Tx, double *Tr, int L, double *S) {
-    for (int k = 0; k < L; k++) {
-        double res  = 1.;
-        double *Trk = Tr + k * L;
-        for (int j = 1; j < L; j++) {
-            res += 2. * Tx[j] * Trk[j];
-        }
-        S[k] = res / (double)(L);
-    }
-}
-inline void part_to_cheb(double x, double y, double z, double *Mu, int Lc) {
-    double *Cx = new double[Lc];
-    double *Cy = new double[Lc];
-    double *Cz = new double[Lc];
-    for (int i = 0; i < Lc; i++) {
-        Cx[i] = C1D(i, x, Lc);
-        Cy[i] = C1D(i, y, Lc);
-        Cz[i] = C1D(i, z, Lc);
-    }
-    for (int i = 0; i < Lc; i++) {
-        for (int j = 0; j < Lc; j++) {
-            for (int k = 0; k < Lc; k++) {
-                Mu[i * Lc * Lc + j * Lc + k] = Cx[i] * Cy[j] * Cz[k];
-            }
-        }
-    }
-}
-
 template <typename CoefficientPrecision, typename CoordinatePrecision, int dimension>
-class IBcomp final : public VirtualInternalLowRankGenerator<CoefficientPrecision> {
+class HCA final : public VirtualInternalLowRankGenerator<CoefficientPrecision> {
 
   private:
+    std::function<void(std::array<CoordinatePrecision, dimension> *, int, std::array<CoordinatePrecision, dimension> *, int, CoefficientPrecision *)> m_kernel;
     const int *m_target_permutation;
     const int *m_source_permutation;
-    std::function<CoefficientPrecision(CoordinatePrecision *, CoordinatePrecision *)> m_kernel;
-    const std::map<int, std::vector<std::pair<int, CoordinatePrecision>>> &
-        m_target_dof_to_elts;
-    const std::map<int, std::vector<std::pair<int, CoordinatePrecision>>> &m_source_dof_to_elts;
+    const CoordinatePrecision *m_target_points;
+    const CoordinatePrecision *m_source_points;
[…]
