+++ CMakeLists.txt
@@ -78,6 +78,7 @@ endif()
 option(HTOOL_WITH_EXAMPLES "Build htool examples ?" ON)
 option(HTOOL_WITH_DOC "Build documentation" ON)
 option(HTOOL_WITH_STRICT_TESTS "Add -Werror to the tests" OFF)
+option(HTOOL_WITH_THEIA "Add Theia library" ON)
 
 #=============================================================================#
 #========================== External Libraries ===============================#
@@ -123,6 +124,25 @@ message(STATUS "Lapack implementation:" "${LAPACK_LIBRARIES}")
 # HPDDM
 find_package(HPDDM)
 
+# Theia
+if(HTOOL_WITH_THEIA)
+    include(FetchContent)
+    FetchContent_Declare(
+        theia
+        GIT_REPOSITORY https://github.com/IChollet/theia_dev
+        GIT_TAG 563f97f36ef7c79306dd7ba32aba25fd15d3e0e5
+        SOURCE_SUBDIR "MADE-UP-DIRECTORY") # https://stackoverflow.com/questions/79261625/cmake-fetchcontent-with-header-only-project
+
+    FetchContent_MakeAvailable(theia)
+
+    add_library(theia INTERFACE)
+    target_include_directories(theia INTERFACE $<BUILD_INTERFACE:${theia_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>)
+
+    install(
+        TARGETS theia
+        EXPORT ${PROJECT_NAME}Targets
+        PUBLIC_HEADER DESTINATION include)
+endif()
 #=============================================================================#
 #========================== Libraries ========================================#
 #=============================================================================#
@@ -134,6 +154,10 @@ if(OpenMP_CXX_FOUND)
     target_link_libraries(htool INTERFACE OpenMP::OpenMP_CXX)
 endif()
 target_compile_features(htool INTERFACE cxx_std_17)
+
+if(HTOOL_WITH_THEIA)
+    target_link_libraries(htool INTERFACE theia)
+endif()
 
 if("${BLA_VENDOR}" STREQUAL "Intel10_32"
    OR "${BLA_VENDOR}" STREQUAL "Intel10_64lp"
