### 2026-02-17 #9279 @PierreMarchand20
Dear @gkthiruvathukal, is everything ok concerning the paper?
