### 2026-02-17 6bcb49e
minor changes in paper.md
