+++ paper.md
@@ -34,7 +34,7 @@ Compressible dense linear systems arise in many applications such as:
 
 - discretization of boundary integral equations [@BoermGrasedyckEtAl2003IHM],
 - solving Lyapunov and Riccati equations [@BoermGrasedyckEtAl2003IHM],
-- discretization of the integral Fractional Laplacian [@AinsworthGlusa2018TEF],
+- discretization of the integral Fractional Laplacian [@AinsworthGlusa2018TEF], and
 - kernel-based scattered data interpolation [@IskeLeBorneEtAl2017HMA].
 
 They are often derived from a discretization of asymptotically smooth kernels $\kappa(\mathbf{x},\mathbf{y})$, i.e., for $\mathbf{x}\neq\mathbf{y}$
@@ -45,22 +45,22 @@ where $\alpha,\beta \in \mathbb{N}_0^d$, with $\alpha+\beta\neq 0$ and $C_{\math
 
 This mathematical property implies that the interaction between two distant clusters can be represented by a low-rank matrix. Multiple compression techniques have been developed to take advantage of this low-rank structure and provide an approximated representation of the a priori dense linear systems, for example, $\mathcal{H}$-matrix and $\mathcal{H}^2$-matrix [@Hackbusch2015HMA; @Bebendorf2008HMM; @Borm2010ENM] and FMM [@GreengardRokhlin1987FAP]. The advantages are usually:
 
-- to avoid the assembly of the full dense linear system,
+- to avoid the assembly of the full dense linear system, and
 - to provide a compressed/approximated version of all or part of the usual dense linear algebra.
 
 The goal of such compression techniques is to lower the memory footprint and the cost of solving the linear system. For the latter, they provide an approximated matrix-vector product whose complexity scales linearly or quasi-linearly with the size of the problem. Thus, iterative solvers such as Conjugate Gradient or GMRES are well-suited.
 
 Beside compression, another technique to accelerate iterative linear solvers is to use preconditioners: instead of solving $\mathbf{A}\mathbf{x}=\mathbf{f}$, where $\mathbf{A}$ is a dense or compressed linear system, we solve $\mathbf{P}\mathbf{A}\mathbf{x}=\mathbf{P}\mathbf{f}$ where $\mathbf{P}$ should be relatively inexpensive to apply, and $\mathbf{P}\mathbf{A}$ more suitable for an iterative resolution.
 
-A class of preconditioners stemming from domain decomposition methods (DDM) are *Schwarz preconditioners*. They rely on a decomposition with overlap of the initial domain. $\mathbf{P}$ is constructed by solving well-chosen problems within each subdomain, together with a small global problem, called coarse space, that provides the necessary global information for the method to retain robustness as the number of subdomains increases. Such preconditioners have been introduced for boundary integral equations in [@Hebeker1990PSA] and analyzed in [@StephanTran1998DDA; @Heuer1996EA].
+A class of preconditioners stemming from domain decomposition methods (DDM) are *Schwarz preconditioners*. They rely on a decomposition with overlap of the initial domain. $\mathbf{P}$ is constructed by solving well-chosen problems within each subdomain, together with a small global problem, called coarse space, that provides the necessary global information for the method to retain robustness as the number of subdomains increases. Such preconditioners have been introduced for boundary integral equations in @Hebeker1990PSA and analyzed in @StephanTran1998DDA and @Heuer1996EA.
 
[…]
