### 2026-02-17 bec668b
minor changes in paper.bib
