+++ paper.bib
@@ -133,7 +133,7 @@
   pages     = {17--57},
   editor    = {Dick, Josef and Kuo, Frances Y. and Wo{\'{z}}niakowski, Henryk},
   publisher = {Springer International Publishing},
-  title     = {Towards an Efficient Finite Element Method for the Integral Fractional Laplacian on Polygonal Domains},
+  title     = {Towards an Efficient Finite Element Method for the Integral Fractional {L}aplacian on Polygonal Domains},
   year      = {2018},
   isbn      = {9783319724560},
   booktitle = {Contemporary Computational Mathematics - A Celebration of the 80th Birthday of {Ian Sloan}},
@@ -160,7 +160,7 @@
 @inproceedings{Hebeker1990PSA,
   author       = {Friedrich K. Hebeker},
   booktitle    = {Domain Decomposition Methods for Partial Differential Equations},
-  title        = {On a parallel Schwarz algorithm for symmetric strongly elliptic integral equations},
+  title        = {On a parallel {S}chwarz algorithm for symmetric strongly elliptic integral equations},
   year         = {1990},
   editor       = {Roland Glowinski and Yuri A. Kuznetsov and G\'erard Meurant and Jacques P\'eriaux and Olaf B. Widlund},
   organization = {SIAM},
