### 2026-02-17 #9279 @danielskatz
@PierreMarchand20 - I've proofread this submission and suggest the changes in https://github.com/htool-ddm/htool/pull/79 - please merge this or let me know what you disagree with, then we can proceed to acceptance and publication
