### 2026-02-17 #9279 @danielskatz
@PierreMarchand20 - please check the proof this will generate and make sure it looks correct. Then we can finish this up.
