### 2026-02-17 #9279 @PierreMarchand20
I checked the proofs with my co-authors, and everything looks good, thank you.
