### 2026-02-17 #9279 @danielskatz
Congratulations to @PierreMarchand20 (Pierre Marchand) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you.
