### 2026-02-18 #9279 @PierreMarchand20
@danielskatz I just noticed there is dot at the end of the title... would it be possible to remove it at least from the metadata? Otherwise it is a big deal, sorry about that...
