### 2026-02-18 #9279 @danielskatz
Please update the paper.md file and ping me - I can then reaccept the paper. Sorry for missing this...
