### 2026-02-18 #9279 @PierreMarchand20
I updated paper.md. Sorry again, thank you for your help 🙏🙏🙏
